* Upgrade Gradle to latest version * Upgrade Android Studio to latest version (Arctic Cat) * Add initial workflow support * Create DurationUtil helper class * Create WorkflowListActivity and WorkflowsFragment (prefer fragment right now)

+++ app/src/main/java/com/jeremyhahn/cropdroid/config/ChannelParser.kt
@@ -27,7 +27,7 @@ class ChannelParser {
         fun parse(jsonChannel: JSONObject) : Channel {
             val id = jsonChannel.getLong("id")
             val controllerId =  if(jsonChannel.isNull("device_id")) jsonChannel.getLong("deviceId") else jsonChannel.getLong("device_id")
-            val channelId = if(jsonChannel.isNull("channel_id")) jsonChannel.getInt("channelId") else jsonChannel.getInt("channel_id")
+            val channelId = if(jsonChannel.isNull("channel_id")) jsonChannel.getLong("channelId") else jsonChannel.getLong("channel_id")
             val name = jsonChannel.getString("name")
             val enable = jsonChannel.getBoolean("enable")
             val notify = jsonChannel.getBoolean("notify")
